'правки_pt2'
